+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+* Reset all camping time trackers after advancing 24 hours or more and not just the one from the active party
+
 ## [5.9.1] - 2026-01-11
 
 ### Changed